Extract code at build time
