+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "bundlewasm": "node scripts/bundlewasm.js",
+    "bundlewasm": "ts-node --esm scripts/bundlewasm.ts",
     "check": "tsc --noEmit",
     "format": "prettier --write --no-semi .",
     "preview": "vite preview",
@@ -17,6 +17,7 @@
   },
   "devDependencies": {
     "assemblyscript": "^0.27.11",
+    "ts-node": "^10.9.1",
     "typescript": "^5.0.2",
     "vite": "^4.4.5",
     "vitest": "^0.34.5"
